# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para la gestión de una peluquería",
        "title": "Peluquería API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/producto": {
            "get": {
                "description": "Devuelve una lista de todos los productos registrados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dtos.GetProductDto"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo producto en el sistema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "parameters": [
                    {
                        "description": "Datos del producto",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateProductDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/producto/{id}": {
            "get": {
                "description": "Devuelve los datos de un producto específico.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener producto por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.GetProductDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza los datos de un producto específico.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del producto",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateProductDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos o ID inválidos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un producto específico.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/producto/{id}/restock": {
            "post": {
                "description": "Agrega stock adicional a un producto existente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Productos"
                ],
                "summary": "Reabastecer producto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos para el reestock",
                        "name": "restock",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RestockProductDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos o ID inválidos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CreateProductDto": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "Head \u0026 Shoulders"
                },
                "low_stock_alert": {
                    "type": "number",
                    "example": 100
                },
                "name": {
                    "type": "string",
                    "example": "Shampoo Anticaspa"
                },
                "package_count": {
                    "type": "number",
                    "example": 32
                },
                "unit": {
                    "type": "string",
                    "example": "ml"
                },
                "unit_per_package": {
                    "type": "number",
                    "example": 500
                },
                "unity_price": {
                    "type": "number",
                    "example": 10000
                }
            }
        },
...
        "dtos.GetProductDto": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "Head \u0026 Shoulders"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Shampoo Anticaspa"
                },
                "quantity": {
                    "type": "number",
                    "example": 400
                },
                "unit": {
                    "type": "string",
                    "example": "ml"
                }
            }
        },
...
        "dtos.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Datos adicionales"
                },
                "message": {
                    "description": "Mensaje descriptivo",
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "status": {
                    "description": "Ejemplo de estado para éxito o error",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "dtos.RestockProductDto": {
            "type": "object",
            "properties": {
                "package_count": {
                    "type": "number",
                    "example": 32
                },
                "reason": {
                    "description": "Razón del movimiento (opcional)",
                    "type": "string",
                    "example": "Reinventario"
                },
                "unit_per_package": {
                    "type": "number",
                    "example": 500
                },
                "unity_price": {
                    "description": "Precio unitario",
                    "type": "number",
                    "example": 10000
                }
            }
        },
...
        "dtos.UpdateProductDto": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "Head \u0026 Shoulders"
                },
                "low_stock_alert": {
                    "type": "number",
                    "example": 100
                },
                "name": {
                    "type": "string",
                    "example": "Shampoo Anticaspa"
                },
                "unit": {
                    "type": "string",
                    "example": "ml"
                }
            }
        },
...
    }
}